feat: Add Github Actions CI workflow for Java

+++ src/main/java/org/example/demosof3041/ontap1/TbcMang.java
@@ -9,6 +9,6 @@ public class TbcMang {
         for(int value : arr){
             tong += value;
         }
-        return (int) tong / arr.length;
+        return tong / arr.length;
     }
 }
